requirements.txt
langchain-openai
langchain-experimental
google-api-python-client
google-auth-oauthlib
google-genai
numpy
protobuf
Pillow==11.3.0
moviepy==2.2.1